lms_log_analyzer/requirements.txt
faiss-cpu
langchain-google-genai
langchain-core
sentence-transformers
google-api-python-client
requests
fastapi
uvicorn
python-grok
py2neo
